홈화면 response 수정

+++ src/main/java/com/dbproject/startupspace/repository/ProvincialCenterEntityRepository.java
@@ -1,8 +1,9 @@
 package com.dbproject.startupspace.repository;
 
+import com.dbproject.startupspace.domain.dto.CenterDTO;
 import com.dbproject.startupspace.domain.entity.MetropolitanCenterEntity;
 import com.dbproject.startupspace.domain.entity.ProvincialCenterEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvincialCenterEntityRepository extends JpaRepository<ProvincialCenterEntity, Long> {
+public interface ProvincialCenterEntityRepository extends JpaRepository<ProvincialCenterEntity, Integer> {
 }
